+++ src/core/memory/pool_allocator.cpp
@@ -72,7 +72,7 @@ void *PoolAllocator::allocate(u64 size, u64 alignment) {
 
   Header *header = reinterpret_cast<Header *>(headerPtr);
   header->size = size;
-  header->alignment = alignment;
+  header->alignment = static_cast<u32>(alignment);
   header->allocationId = AllocCounter::add();
   header->allocatorId = AllocatorType::pool;
   header->state = HeaderState::allocated;
