+++ src/core/memory/stack_allocator.cpp
@@ -49,7 +49,7 @@ void *StackAllocator::allocate(u64 size, u64 alignment) {
 
   Header *header = reinterpret_cast<Header *>(headerPtr);
   header->size = size;
-  header->alignment = alignment;
+  header->alignment = static_cast<u32>(alignment);
   header->allocationId = AllocCounter::add();
   header->allocatorId = AllocatorType::stack;
   header->state = HeaderState::allocated;
